refactor(WinningRankCalculator): 모듈 및 메소드 이름 직관적으로 변경
